% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_autoSurv.R
\name{predict_autoSurv}
\alias{predict_autoSurv}
\title{Discrete-time survival prediction models}
\usage{
predict_autoSurv(object, newdata, times, timeVar, statusVar)
}
\arguments{
\item{object}{autoSurv() object that contains trained models}

\item{newdata}{data set on which you want to obtain survival predictions}

\item{times}{vector of time points at which survival probability is predicted}

\item{timeVar}{string corresponding to the variable name of the time-to-event outcome. Used to compute performance metrics}

\item{statusVar}{string corresponding to the variable name of the status indicator. Used to compute performance metrics}
}
\value{
A list with predicted probabilities and performance metrics
}
\description{
Compute the predicted survival probabilities for trained discrete-time models
}
\section{Authors}{

Krithika Suresh (\email{krihtika.suresh@cuanschutz.edu})
}

