% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSurv.R
\name{autoSurv}
\alias{autoSurv}
\title{autoSurv}
\usage{
autoSurv(
  timeVar,
  statusVar,
  data,
  times = NULL,
  trainModels = c("cox", "rsf", "glm", "gbm", "glmnet", "svm", "cforest", "nnet"),
  bins = 10,
  testProp = 0.2,
  seed = 1,
  init = 10,
  iter = 20,
  cv = TRUE,
  cvFold = 5
)
}
\arguments{
\item{timeVar}{string corresponding to the variable name of the time-to-event}

\item{statusVar}{string corresponding to the variable name of the status indicator}

\item{data}{data frame containing covariates for prediction}

\item{times}{vector of time points at which probably of event is predicted}

\item{trainModels}{list of models to train. options: "cox","rsf","glm","gam","gbm","glmnet","svm","cforest","nnet"}

\item{bins}{number of time bins}

\item{testProp}{proportion of observations to be randomly selected for the test set}

\item{seed}{integer random seed}

\item{init}{number of randomly chosen points to sample the target function before Bayesian Optimization fitting the Gaussian Process}

\item{iter}{total number of times the Bayesian Optimization is to be repeated}

\item{cv}{enable or disable cross-validation}

\item{cvFold}{number of folds to use in k-fold cross validation}
}
\value{

}
\description{
autoSurv
}
